Adjust Primary Navigation.  Details below...
Search is the default action for the homepage, so we've introduced a search
input to recieve queries on this page rather than redirecting the user
elsewhere before they can do something useful.

The search form needs to match the default search form more closely so that
moving between the two isn't as jarring.  This might include adding a collection
specifier as this might be a contentious issue.  Namely, not being able to resrict the collection might irritate users of the smaller collections.

We've shortened 'About Dippam' just to 'About' since it sits directly under the
Dippam title and seems redundant.

We've hidden the collection listings under a 'Collections' menu item, since having them all listed isn't of interest to return users and new users might be overwhelmed by the number of items.  They are listed in full in the centre of the page in any case and still available under the 'Collections' label.